Move all classes into Ingress.java
The kamel cli can't resolve
dependencies between class files
so I dumped everything into a
single file for the time being
until I can get the route to
actually respond to a message.
