--- camel-k/src/main/java/io/ppatierno/formula1/PC2CamelKAppConfig.java
@@ -1,37 +1,0 @@
-/*
- * Copyright Paolo Patierno.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.ppatierno.formula1;
-
-public class PC2CamelKAppConfig {
-
-    private static final String UDP_PORT_ENV = "UDP_PORT";
-
-    private static final int DEFAULT_UDP_PORT = 10666;
-
-    private final int udpPort;
-
-    private PC2CamelKAppConfig(int udpPort) {
-        this.udpPort = udpPort;
-    }
-
-    public static PC2CamelKAppConfig fromEnv() {
-        int udpPort = System.getenv(UDP_PORT_ENV) == null
-            ? DEFAULT_UDP_PORT
-            : Integer.parseInt(System.getenv(UDP_PORT_ENV));
-
-        return new PC2CamelKAppConfig(udpPort);
-    }
-
-    public int getUdpPort() {
-        return udpPort;
-    }
-
-    @Override
-    public String toString() {
-        return "F1CamelKAppConfig[" +
-            "udpReceivePort=" + this.udpPort +
-            "]";
-    }
-}
