--- camel-k/src/main/java/io/ppatierno/formula1/pc2/Terrain.java
@@ -1,19 +1,0 @@
-/**
- * Terrain definition (based on code from https://github.com/ralfhergert/pc2-telemetry)
- */
-
-package io.ppatierno.formula1.pc2;
-
-/**
- * Defines all terrains.
- * Don't change the order. The ordinal is important.
- */
-public enum Terrain {
-	Asphalt,
-	u1,
-	u2,
-	u3,
-	u4,
-	u5,
-	Grass
-}
