--- camel-k/src/main/java/io/ppatierno/formula1/pc2/Tire.java
@@ -1,16 +1,0 @@
-/**
- * Tire definition (based on code from https://github.com/ralfhergert/pc2-telemetry)
- */
-
-package io.ppatierno.formula1.pc2;
-
-/**
- * This enum defines the tyres.
- * Don't change the order. The ordinal is important.
- */
-public enum Tire {
-	FrontLeft,
-	FrontRight,
-	RearLeft,
-	RearRight,
-}
